refactor: move EmailValidatorAdapter to the main layer
