+++ src/main/adapters/validators/email-validator-adapter.ts
@@ -1,5 +1,5 @@
 import validator from 'validator'
-import { EmailValidator } from '../presentation/protocols/email-validator'
+import { EmailValidator } from '../../../presentation/protocols/email-validator'
 
 export class EmailValidatorAdapter implements EmailValidator {
   isValid(email: string): boolean {
