+++ src/main/factories/login/login-validation-factory.ts
@@ -4,7 +4,7 @@ import {
   RequiredFieldValidation,
 } from '../../../presentation/helpers/validators'
 import { Validation } from '../../../presentation/protocols/validation'
-import { EmailValidatorAdapter } from '../../../utils/email-validator-adapter'
+import { EmailValidatorAdapter } from '../../adapters/validators/email-validator-adapter'
 
 export const makeLoginValidation = (): ValidationComposite => {
   const requiredFields: Validation[] = []
