+++ src/main/factories/signup/signup-validation-factory.ts
@@ -5,7 +5,7 @@ import {
   CompareFieldsValidation,
 } from '../../../presentation/helpers/validators'
 import { Validation } from '../../../presentation/protocols/validation'
-import { EmailValidatorAdapter } from '../../../utils/email-validator-adapter'
+import { EmailValidatorAdapter } from '../../adapters/validators/email-validator-adapter'
 
 export const makeSignUpValidation = (): ValidationComposite => {
   const requiredFields: Validation[] = []
